Add LinkedList and LinkedDictionary
I’m thinking of implementing the objects representing signal paths
using some kind of ordered dictionary.

This way, the components could be iterated over (e.g. for checking what
controls are available), but could also be accessed directly.

I’ve made a first draft that uses a linked list to maintain the order
of the items, and then a dictionary mapping component names to the list
nodes the components are contained in.

The problem with this is that if you iterate over the nodes, you can’t
see the keys in the dictionary, only the values.