Add Keys as Properties on Linked List Nodes
I’m not sure if this is the most sensible solution, but I’ve decided to
allow access to the keys by adding them to the linked list nodes. I’m
using the NamedListNode object for the linked list nodes, which has an
extra getName/setName function pair.

I also refactored the add methods, separating the commonalities into
one method add(), and added an iterator over the keys.