Create ConfigurableObject
I want things I create, starting with the Sampler, to be very
configurable by changing object literals or adding new ones, or
possibly using JSON.

To aid this, I’ve created ConfigurableObject, which provides the
functionality required for an object to be configurable.

* configure() is a method which carries out the configuration using the
parameters it’s passed

* setConfigurableParameters() is a method used to set up the
configuration options that are available on an object

I’ve also tested the process with newObject.

It’s pretty early yet, but I like this a lot.