Change samples
I’ve realised that using the samples I was using was copyright
infringement, so I’ve chosen some new samples that should be fine.

It’d be nice to make the sampler read in sample choices and positioning
from a file in future, rather than hardcoding them into the javascript.